Make the jvm interop layer support --option
We now have overrideable defaults via the '-o' and '--option' args.

As part of this work, I decided to fork org.clojure/tools.cli and bundle
a copy of it with streamparse. Why? Classpath hell. The lein exclusions
thing just wasn't working correctly in the case of project dependencies.
Thanksfully, it's just a couple hundred lines of code and a single file,
so now it lives in `streamparse.cli` and I updated our tools to use it.

I needed the newer version of tools.cli to get the :assoc-fn
functionality that was missing in Storm's bundled version.