Fix issue #21.
It turns out that logger.error() which we used to use in LogStream
will call sys.stderr under the hood which we've redirected to
LogStream which creates an infinite recursion loop.

Solution here was to send stderr/stdout attempts to logger.info().
When we add support for logging in streamparse, we'll be hoping that
users use a FileHandler for all logging which will then capture all
print statements correctly.

We're also putting in a safety check in the event something goes
really wrong with logging, we just raise an exception to Storm which
will kill the worker process.