Major test case refactor.
DRY test cases via ShellComponentTestCaseMixin which now takes care of
creating the Python subprocess based on the COMPONENT class var.  Also
handles the initial handshake with the component.

We still have the separate Python file per component, but it’s not so
bad.

Refactored existing tests to use this and added new tests to cover
cases like auto: acking, anchoring and failing.

Was able to find a particularly odd bug which is preventing auto
anchoring working on BatchingBolt. Will need some help to figure that
one out.