Override BatchingBolt.emit
BatchingBolt.emit will not return task IDs a tuple was emitted to as it
currently causes a complication due to the _batcher and main thread
calling read_task_ids and read_command respectively.

Properly returning task IDs will require some major reworking of the
IPC layer.

Since we are sure that only the main thread will be reading from stdin
now, we can move tuple reading and group keys outside of the
_batch_lock.