Reorganized all of the IPC code, so that there's now a "storm" sub-package.
This will eventually turn into the separate "pystorm" package we plan to release.

For now, the unit tests aren't passing, so some things have to be tweaked.  The
main issue is that some of the mocking is no longer necessary, because you can
pass input_stream and output_stream paramamters to the Component initializer for
testing.  They just have to be lists of bytes.