Make BatchingBolt use tick tuples instead of threads.
This still allows for non-time-based batching, it just doesn't require threads
anymore.

This is not a backward-compatible change, as secs_between_branches was changed
to ticks_between_branches, and people now need to run Storm with
topology.tick.tuple.freq.secs set to 1 to have their code work the same as
before.