Fix issue where StopIteration exception was being raised when using next instead of readilne.
This only manifested itself when using a real stdin talking to Storm. I vaguely
recall running into this problem when using next() instead of readline with
stdin in the past, even though it works fine with other files.  Anyway, since
this was really just for testing, I've just made it so we use readline and the
tests now use io.StringIO objects for stdin instead of lists.