Add preliminary support for remote debugging.
You can now send a USR1 signal to a bolt or spout to have it pause execution and
allow you to debug it interactively via telnet.  The port should be logged, but
that does not always seem to work for some reason.  For now just use netstat to
find out the port.  The host is always set to 127.0.0.1 for now, but this will
be more flexible in the future.