upload jar before killing existing topology. (#277)
if we kill the topology before uploading the jar, the topology stays disabled until the upload is complete. on a slow network, this can lead to the topology being disabled for an unacceptably long time during the deployment. additionally, if the upload fails and deployment crashes, this change will ensure that it did not have a destructive effect on the existing topology.